add depot map and documentation
