+++ c300/c300_description/urdf/c300.ros2_control.xacro
@@ -4,40 +4,44 @@
   <xacro:macro name="c300_ros2_control" params="
     name
     use_fake_hardware:=false
+    sim_isaac:=false
     sim_ignition:=false">
 
-    <xacro:if value="${sim_ignition}">
-      <ros2_control name="${name}" type="system">
-        <hardware>
-          <xacro:if value="${sim_ignition}">
-            <plugin>ign_ros2_control/IgnitionSystem</plugin>
-          </xacro:if>
-          <xacro:if value="${use_fake_hardware}">
-            <plugin>fake_components/GenericSystem</plugin>
-            <param name="fake_sensor_commands">${fake_sensor_commands}</param>
-            <param name="state_following_offset">0.0</param>
-          </xacro:if>
-        </hardware>
-        <joint name="drivewhl_l_joint">
-          <command_interface name="velocity">
-            <param name="min">-3.15</param>
-            <param name="max">3.15</param>
-          </command_interface>
-          <state_interface name="position"/>
-          <state_interface name="velocity"/>
-          <state_interface name="effort"/>
-        </joint>
-        <joint name="drivewhl_r_joint">
-          <command_interface name="velocity">
-            <param name="min">-3.15</param>
-            <param name="max">3.15</param>
-          </command_interface>
-          <state_interface name="position"/>
-          <state_interface name="velocity"/>
-          <state_interface name="effort"/>
-        </joint>
-      </ros2_control>
-    </xacro:if>
+    <ros2_control name="${name}" type="system">
+      <hardware>
+        <xacro:if value="${sim_ignition}">
+          <plugin>ign_ros2_control/IgnitionSystem</plugin>
+        </xacro:if>
+        <xacro:if value="${sim_isaac}">
+          <plugin>topic_based_ros2_control/TopicBasedSystem</plugin>
+          <param name="joint_commands_topic">/isaac_joint_commands</param>
+          <param name="joint_states_topic">/isaac_joint_states</param>
+        </xacro:if>
+        <xacro:if value="${use_fake_hardware}">
+          <plugin>fake_components/GenericSystem</plugin>
+          <param name="fake_sensor_commands">${fake_sensor_commands}</param>
+          <param name="state_following_offset">0.0</param>
+        </xacro:if>
+      </hardware>
+      <joint name="drivewhl_l_joint">
+        <command_interface name="velocity">
+          <param name="min">-3.15</param>
+          <param name="max">3.15</param>
+        </command_interface>
+        <state_interface name="position"/>
+        <state_interface name="velocity"/>
+        <state_interface name="effort"/>
+      </joint>
+      <joint name="drivewhl_r_joint">
+        <command_interface name="velocity">
+          <param name="min">-3.15</param>
+          <param name="max">3.15</param>
+        </command_interface>
+        <state_interface name="position"/>
+        <state_interface name="velocity"/>
+        <state_interface name="effort"/>
+      </joint>
+    </ros2_control>
   </xacro:macro>
 
 </robot>
