+++ c300/c300_description/urdf/c300_base.urdf.xacro
@@ -1,17 +1,23 @@
 <?xml version="1.0"?>
 <robot name="c300" xmlns:xacro="http://ros.org/wiki/xacro">
 
-  <xacro:macro name="load_c300" params="prefix sim_ignition:=false use_fake_hardware:=false">
+  <xacro:macro name="load_c300" params="
+    prefix
+    sim_isaac:=false
+    sim_ignition:=false
+    use_fake_hardware:=false">
+
     <!-- Import macros for main hardware components -->
     <xacro:include filename="$(find c300_description)/urdf/hokuyo_uam05lp.urdf.xacro" />
 
-    <xacro:if value="${use_fake_hardware or sim_ignition}">
+    <xacro:if value="${use_fake_hardware or sim_ignition or sim_isaac}">
       <!-- ros2 control is only used for simulation -->
       <xacro:include filename="$(find c300_description)/urdf/c300.ros2_control.xacro" />
       <xacro:property name="ros2_control_name" value="IgnitionSystem"/>
       <xacro:c300_ros2_control
         name="${ros2_control_name}"
         use_fake_hardware="${use_fake_hardware}"
+        sim_isaac="${sim_isaac}"
         sim_ignition="${sim_ignition}"/>
       <xacro:include filename="$(find c300_description)/urdf/c300_transmission_macro.xacro" />
       <xacro:load_gazebo_base_transmissions/>
