+++ c300/c300_navigation/map/depot.yaml
@@ -1,0 +1,7 @@
+image: depot.pgm
+mode: trinary
+resolution: 0.05
+origin: [-15.2, -7.72, 0]
+negate: 0
+occupied_thresh: 0.65
+free_thresh: 0.25
