--- c300/c300_navigation/map/empty_world.yaml
@@ -1,6 +1,0 @@
-image: empty_world.pgm
-resolution: 0.050000
-origin: [-1024.000000, -1024.000000, 0.000000]
-negate: 0
-occupied_thresh: 0.65
-free_thresh: 0.196
