--- c300/c300_navigation/map/house.yaml
@@ -1,7 +1,0 @@
-image: house.pgm
-mode: trinary
-resolution: 0.05
-origin: [-9.29, -5.92, 0]
-negate: 0
-occupied_thresh: 0.65
-free_thresh: 0.25
